+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 2.5.0
 -----
 
+ * Added `translation:debug` command
  * Added `config:debug` command
  * Added `yaml:lint` command
  * Deprecated the `RouterApacheDumperCommand` which will be removed in Symfony 3.0.